refactor(texas): Move MergeResult and docket disaggregation
