+++ cl/corpus_importer/tasks.py
@@ -8,7 +8,6 @@ import re
 import shutil
 from collections.abc import Iterator
 from contextlib import contextmanager
-from dataclasses import dataclass, field
 from datetime import date
 from http import HTTPStatus
 from io import BytesIO
@@ -117,6 +116,7 @@ from cl.corpus_importer.llm_models import CaseNameExtractionResponse
 from cl.corpus_importer.management.utils import TexasDocketMeta
 from cl.corpus_importer.prompts.system import CASE_NAME_EXTRACT_SYSTEM
 from cl.corpus_importer.state.texas.utils import is_missing_file_page
+from cl.corpus_importer.state.utils import MergeResult
 from cl.corpus_importer.utils import (
     DownloadPDFResult,
     compute_binary_probe_jitter,
@@ -178,6 +178,7 @@ from cl.recap.mergers import (
     add_claims_to_docket,
     add_parties_and_attorneys,
     add_tags_to_objs,
+    find_and_disaggregate_docket_object,
     find_docket_object,
     make_recap_sequence_number,
     merge_pacer_docket_into_cl_docket,
@@ -4054,90 +4055,6 @@ def download_texas_document_unthrottled(
 TAMES_PENDING_SUBSCRIPTIONS_KEY = "tames:pending_subscriptions"
 
 
-@dataclass
-class MergeResult[T = int]:
-    """Stores data about the result of an attempted merge operation.
-
-    :ivar creates: Objects which needed to be created. Key is object name and
-        value is a list of PKs to created objects.
-    :ivar updates: Objects which needed to be updated.
-    :ivar failures: Objects for which the merge operation failed. Items will be
-        None if an object needed to be created but that operation failed."""
-
-    creates: dict[str, set[T]] = field(default_factory=dict)
-    updates: dict[str, set[T]] = field(default_factory=dict)
-    failures: dict[str, list[T | None]] = field(default_factory=dict)
-
-    @staticmethod
-    def union[S, U](
-        a: MergeResult[S], b: MergeResult[U]
-    ) -> MergeResult[S | U]:
-        """
-        Creates a new MergeResult object storing the combined results of two
-        objects.
-        """
-        return MergeResult[S | U](
-            creates={
-                k: a.creates.get(k, set()) | b.creates.get(k, set())
-                for k in a.creates.keys() | b.creates.keys()
-            },
-            updates={
-                k: a.updates.get(k, set()) | b.updates.get(k, set())
-                for k in a.updates.keys() | b.updates.keys()
-            },
-            failures={
-                k: [*a.failures.get(k, []), *b.failures.get(k, [])]
-                for k in a.failures.keys() | b.failures.keys()
-            },
-        )
-
-    @property
-    def success(self) -> bool:
-        return not self.failures
-
-    @property
-    def update(self) -> bool:
-        return bool(self.updates)
-
-    @property
-    def create(self) -> bool:
-        return bool(self.creates)
-
-    @staticmethod
-    def created[S](model: str, pk: S) -> MergeResult[S]:
-        """Shorthand for the result of a successful create operation.
-
-        :param model: The model which was created.
-        :param pk: The primary key of created object.
-        :returns: The constructed MergeResult object."""
-        return MergeResult(creates={model: {pk}})
-
-    @staticmethod
-    def updated[S](model: str, pk: S) -> MergeResult[S]:
-        """Shorthand for the result of a successful update operation.
-
-        :param model: The model which was updated.
-        :param pk: The primary key of the updated object.
-        :return: The constructed MergeResult object."""
-        return MergeResult(updates={model: {pk}})
-
-    @staticmethod
-    def failed[S](model: str, pk: S | None = None) -> MergeResult[S]:
-        """Shorthand for the result of a failed merge operation.
-
-        :param model: The model which failed.
-        :param pk: The (optional) primary key of the failed object.
-        :return: The constructed MergeResult object."""
-        return MergeResult(failures={model: [pk]})
-
-    @staticmethod
-    def unnecessary() -> MergeResult:
-        """Shorthand for the result of an unnecessary merge operation.
-
-        :return: The constructed MergeResult object."""
-        return MergeResult()
-
-
 def merge_texas_trial_court_data(
     docket: Docket,
     docket_data: TexasCourtOfCriminalAppealsDocket | TexasSupremeCourtDocket,
@@ -4880,14 +4797,19 @@ def merge_texas_docket(
     :param download_attachments: Whether to download docket entry attachments.
 
     :return: The result of the merge operation."""
-    court = Court.objects.get(
-        pk=texas_js_court_id_to_court_id(docket_data["court_id"])
-    )
+    court_pk = texas_js_court_id_to_court_id(docket_data["court_id"])
     docket_number = docket_data["docket_number"]
     logger.info("Merging Texas docket %s", docket_number)
 
     if docket_data["court_type"] == CourtType.UNKNOWN.value:
         logger.error("Texas docket %s has unknown court type", docket_number)
+        return MergeResult.failed("Docket")
+    if court_pk is None:
+        logger.error(
+            "Could not determine Court pk for Texas docket %s with court ID %s",
+            docket_number,
+            docket_data["court_id"],
+        )
         return MergeResult.failed("Docket")
 
     with transaction.atomic():
@@ -4896,24 +4818,21 @@ def merge_texas_docket(
             logger.info(
                 "Docket is appellate. Checking if disaggregation is necessary..."
             )
-            docket = async_to_sync(find_docket_object)(
-                court_id="texapp",
-                pacer_case_id=None,
+            docket, changed = async_to_sync(
+                find_and_disaggregate_docket_object
+            )(
+                court_id=court_pk,
+                aggregate_court_id="texapp",
                 docket_number=docket_number,
-                federal_defendant_number=None,
-                federal_dn_judge_initials_assigned=None,
-                federal_dn_judge_initials_referred=None,
                 docket_source=Docket.SCRAPER,
-                allow_create=False,
             )
-            if docket is not None:
+            if changed:
                 logger.info(
-                    "Disaggregating Texas appellate docket %s", docket_number
+                    f"Disaggregated Texas appellate docket {docket_number}"
                 )
-                docket.court = court
-        if docket is None:
+        else:
             docket = async_to_sync(find_docket_object)(
-                court_id=court.pk,
+                court_id=court_pk,
                 pacer_case_id=None,
                 docket_number=docket_number,
                 federal_defendant_number=None,
